data-deleter/requirements.txt
google.cloud
google.cloud.storage
google.cloud.bigquery
google.api_core
functions-framework==3.2.0